Added Up Horns, Resaturated Textures, and Minimal Animation Code
Animation code needs to be smoothed to not instantly switch poses, thats why the lunge looks weird.  The potion animation code needs to be turned on for maybe a second before turning itself off, currently it kind of works with health potion but the others are broken.  Reloading animation code just doesn't seem to work at all.